requirements.txt
streamlit 
langchain 
openai 
pandas
python-dotenv
langchain-community
faiss-cpu 